Rename the 'msg' to 'mix'
