+++ transport/protocol/benches/protocol_throughput_emulated.rs
@@ -10,7 +10,7 @@ use hopr_crypto_types::keypairs::Keypair;
 use hopr_internal_types::prelude::*;
 use hopr_network_types::prelude::ResolvedTransportRouting;
 use hopr_primitive_types::prelude::{Balance, BalanceType};
-use hopr_transport_protocol::msg::processor::{MsgSender, PacketInteractionConfig, PacketSendFinalizer};
+use hopr_transport_protocol::mix::processor::{MsgSender, PacketInteractionConfig, PacketSendFinalizer};
 use libp2p::PeerId;
 
 const SAMPLE_SIZE: usize = 20;
