+++ transport/protocol/src/mix/mod.rs
@@ -3,4 +3,4 @@ pub mod packet;
 pub mod processor;
 
 pub use codec::v1::MsgCodec;
-pub const CURRENT_HOPR_MSG_PROTOCOL: &str = "/hopr/msg/1.0.0";
+pub const CURRENT_HOPR_MSG_PROTOCOL: &str = "/hopr/mix/1.0.0";
